requirements.txt
requests
pymongo
groq
python-dotenv